finish ex-262. Unsure about why diagonalize works
I believe it's an elegant solution, and expresses exactly what I had in
mind. On the other hand, I've been going a bit through trial and error
on the last part, more than being systematic.

Still, I believe I've been plenty systematic, and can be satisfied with
the the simplicity of the program.